+++ main.go
@@ -131,6 +131,7 @@ func init() {
 	cmds.ACCOUNT_ADD.FlagSet.StringVar(&cmds.S1, "number", "", "The wallet/account number or identifier.")
 	cmds.ACCOUNT_ADD.FlagSet.StringVar(&cmds.S2, "type", "", "The wallet/account type.")
 	cmds.ACCOUNT_ADD.FlagSet.StringVar(&cmds.S3, "provider", "", "The wallet/account provider (bank/organisation).")
+	cmds.ACCOUNT_ADD.FlagSet.StringVar(&cmds.S4, "holder", "", "The wallet/account holder.")
 
 	cmds.ACCOUNT_REMOVE = cli.NewCommand("remove", &cmds.Help, flag.ExitOnError)
 	cmds.ACCOUNT_REMOVE.Usage = "bh account"